+++ src/projectprototype/Menu.java
@@ -1,0 +1,16 @@
+package projectprototype;
+
+import javax.swing.JPanel;
+
+/**
+ *
+ * @author Matthew
+ */
+public class Menu extends JPanel {
+    
+    public GamePanel panel;
+    
+    public Menu(GamePanel panel) {
+        this.panel = panel;
+    }
+}
